backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6
google-generativeai==0.3.1
geopy==2.4.1
httpx==0.25.1
